[DEBUG] amediacodec: fix output buffer polling by doing it in a thread
This fixes the issue with no active renderer and a video frame filter.
When the demuxer was paused or the end of stream was reached, not all
frames were fed to the video frame filter. This was due to the polling
done in queueInputBuffer() (which was not receiving any more frames) and
in dequeueOutputBuffer() (which was blocked in queue->popBuffer()).