[DEV] record_demuxer: keep up when frames are late
When frames are late, seek to the next sync sample (I frame) that
gives a positive wait time, only if the wait time is less than
twice the average frame output rate (i.e. continue to output late
frames until the seek to the next I frame results in a reasonable
wait time).

Change-Id: Ifacaf8ad4c7318ec61e26045fb85dbd267b1b616
Reviewed-on: https://gerrit-doc.parrot.biz/84222
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Aurelien Barre <aurelien.barre@parrot.com>